the single data loaded for single id

+++ app/src/main/java/sanaebadi/info/movieapp/api/MovieApiInterface.kt
@@ -16,6 +16,6 @@ interface MovieApiInterface {
 
     //details
     @GET("movie/{movie_id}")
-    fun getMovieDetails(@Path("movie_id") movieId: Int): Single<MovieDetails>
+    fun getMovieDetails(@Path("movie_id") movieId: Long): Single<MovieDetails>
 
 }
