+++ app/src/main/java/sanaebadi/info/movieapp/model/MovieDetails.kt
@@ -4,7 +4,7 @@ package sanaebadi.info.movieapp.model
 import com.google.gson.annotations.SerializedName
 
 data class MovieDetails(
-    val budget: Int,
+    val budget: Long,
     val id: Int,
     val overview: String,
     val popularity: Double,
@@ -12,7 +12,7 @@ data class MovieDetails(
     val posterPath: String,
     @SerializedName("release_date")
     val releaseDate: String,
-    val revenue: Int,
+    val revenue: Long,
     val runtime: Int,
     val status: String,
     val tagline: String,
